+++ wepbase/src/main/java/com/wep/common/app/Database/RiderSettlement.java
@@ -17,14 +17,14 @@
 package com.wep.common.app.Database;
 
 public class RiderSettlement {
-	
+
 	// Private variable
 	int iBillNumber, iEmployeeId, iTotalItems, iCustId;
-	float fBillAmount, fDeliveryCharge, fPettyCash, fSettledAmount;
-	
+	double fBillAmount, fDeliveryCharge, fPettyCash, fSettledAmount;
+
 	// Default constructor
 	public RiderSettlement(){
-		
+
 		this.iBillNumber = 0;
 		this.iEmployeeId = 0;
 		this.iTotalItems = 0;
@@ -34,11 +34,11 @@ public class RiderSettlement {
 		this.fSettledAmount = 0;
 		this.iCustId = 0;
 	}
-	
+
 	// Parameterized constructor
 	public RiderSettlement(int BillNumber,int EmployeeId,int TotalItems,
-			float BillAmount,float DeliveryCharge,float PettyCash,float SettledAmount, int CustId){
-			
+						   double BillAmount,double DeliveryCharge,double PettyCash,double SettledAmount, int CustId){
+
 		this.iBillNumber = BillNumber;
 		this.iEmployeeId = EmployeeId;
 		this.iTotalItems = TotalItems;
@@ -46,81 +46,81 @@ public class RiderSettlement {
 		this.fDeliveryCharge = DeliveryCharge;
 		this.fPettyCash = PettyCash;
 		this.fSettledAmount = SettledAmount;
-        this.iCustId = CustId;
+		this.iCustId = CustId;
 	}
-	
+
 	// getting BillNumber
 	public int getBillNumber(){
 		return this.iBillNumber;
 	}
-	
+
 	// getting EmployeeId
 	public int getEmployeeId(){
 		return this.iEmployeeId;
 	}
-	
+
 	// getting TotalItems
 	public int getTotalItems(){
 		return this.iTotalItems;
 	}
-	
+
 	// getting BillAmount
-	public float getBillAmount(){
+	public double getBillAmount(){
 		return this.fBillAmount;
 	}
-	
+
 	// getting DeliveryCharge
-	public float getDeliveryCharge(){
+	public double getDeliveryCharge(){
 		return this.fDeliveryCharge;
 	}
-	
+
 	// getting PettyCash
-	public float getPettyCash(){
+	public double getPettyCash(){
 		return this.fPettyCash;
 	}
-		
+
 	// getting SettledAmount
-	public float getSettledAmount(){
+	public double getSettledAmount(){
 		return this.fSettledAmount;
 	}
 
-    // getting CustId
-    public float getCustId(){
-        return this.iCustId;
-    }
-	
+	// getting CustId
+	public float getCustId(){
+		return this.iCustId;
+	}
+
 	// setting BillNumber
 	public void setBillNumber(int BillNumber){
 		this.iBillNumber = BillNumber;
 	}
-	
+
 	// setting EmployeeId
 	public void setEmployeeId(int EmployeeId){
 		this.iEmployeeId = EmployeeId;
 	}
-	
+
 	// setting TotalItems
 	public void setTotalItems(int TotalItems){
 		this.iTotalItems = TotalItems;
 	}
-	
+
 	// setting BillAmount
-	public void setBillAmount(float BillAmount){
+	public void setBillAmount(double BillAmount){
 		this.fBillAmount = BillAmount;
 	}
-	
+
 	// setting DeliveryCharge
-	public void setDeliveryCharge(float DeliveryCharge){
+	public void setDeliveryCharge(double DeliveryCharge){
 		this.fDeliveryCharge = DeliveryCharge;
 	}
-	
+
 	// setting PettyCash
-	public void setPettyCash(float PettyCash){
+	public void setPettyCash(double PettyCash){
 		this.fPettyCash = PettyCash;
 	}
-			
+
 	// setting SettledAmount
-	public void setSettledAmount(float SettledAmount){
+	public void setSettledAmount(double SettledAmount){
 		this.fSettledAmount = SettledAmount;
 	}
